Coroutine should return a channel instead of a read stream
The read stream might already have ended when the coroutine returns.
We need to convert it to a channel before we exit the block. Alternatively,
we could have paused the read stream (converting it to a channel automatically
resumes it) but that doesn't seem to be very proper to me.