Always merge GeoJSON objects to a collection
Since we are now initializing the merger with distinct metadata,
there is no guarantee anymore that the merger switches to the
correct mode, i.e. a collection if more than one chunk should be
merged. The new default is now that the merger always creates a
collection. As a side effect, this also makes the API more
predictable.