Refactor AttributesIndexer to remove panics
Previously illegal transitions and payloads which should contain
values but didn't would cause a panic through an `unreachable` or
an `unwrap`.
These now lead the AttributesIndexer into the `State::Error` state,
instead of crashing the thread.

Change simplifies handling the Payload. Consumers of
the payload are now responsible for converting the received
string into an appropriate type in exchange for saving complicated matching
where is would otherwise be required.